pyproject.toml
[project]
name = "atomic-brain"
version = "0.1.0"
description = "Local-first, LLM-agnostic knowledge management application."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.89.0",
    "fastapi>=0.136.0",
    "fastembed>=0.8.0",
    "lancedb>=0.30.2",
    "litellm>=1.83.9",
    "logfire>=4.32.1",
    "markitdown>=0.1.5",
    "mcp>=1.27.0",
    "pydantic-ai>=1.5.0",
    "typer>=0.15.0",
    "uvicorn>=0.34.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
]

[tool.pytest.ini_options]
minversion = "7.0"
pythonpath = "."

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
